requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bottleneck=1.3.5=py311hbed6279_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
gurobi=10.0.3=py311_0
intel-openmp=2023.1.0=hdb19cb5_46306
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
ncurses=6.4=h6a678d5_0
numexpr=2.8.7=py311h65dcdc2_0
numpy=1.26.2=py311h08b1b3b_0
numpy-base=1.26.2=py311hf175353_0
openssl=3.0.12=h7f8727e_0
pandas=2.1.4=py311ha02d727_0
pip=23.3.1=py311h06a4308_0
python=3.11.5=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.3.post1=py311h06a4308_0
readline=8.2=h5eee18b_0
scipy=1.11.4=py311h08b1b3b_0
setuptools=68.0.0=py311h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
tqdm=4.65.0=py311h92b7b1e_0
tzdata=2023c=h04d1e81_0
wheel=0.41.2=py311h06a4308_0
xz=5.4.5=h5eee18b_0
zlib=1.2.13=h5eee18b_0
